+++ packages/angular/test/base/src/app/standalone/programmatic-modal/modal/modal.component.ts
@@ -1,0 +1,11 @@
+import { Component } from '@angular/core';
+import { IonButton } from "@ionic/angular/standalone";
+
+@Component({
+  selector: 'app-modal',
+  templateUrl: './modal.component.html',
+  imports: [IonButton],
+  standalone: true,
+})
+export class ModalComponent  {
+}